[JOIN LIST] Function request component without accept/reject functionality
